Toolbar menu mobile view and desktop view; Expiry time fix

+++ Project_files/Auction.Client/src/app/services/local-storage.service.ts
@@ -1,24 +1,33 @@
-import { Inject, Injectable, InjectionToken } from '@angular/core';
+import { Inject, Injectable, InjectionToken, PLATFORM_ID } from '@angular/core';
+import { isPlatformBrowser } from '@angular/common';
 
 @Injectable({
   providedIn: 'root'
 })
 export class LocalStorageService {
 
-  constructor() {}
+  constructor(@Inject(PLATFORM_ID) private platformId: Object) {}
 
   public setItem(key: string, value: string) {
-    localStorage.setItem(key, value);
+    if (isPlatformBrowser(this.platformId))
+      localStorage.setItem(key, value);
+    return null;
   }
 
   public getItem(key: string) {
-    return localStorage.getItem(key);
+    if (isPlatformBrowser(this.platformId))
+      return localStorage.getItem(key);
+    return null;
   }
   public removeItem(key: string) {
-    localStorage.removeItem(key);
+    if (isPlatformBrowser(this.platformId))
+      localStorage.removeItem(key);
+    return null;
   }
 
   public clear() {
-    localStorage.clear();
+    if (isPlatformBrowser(this.platformId))
+      localStorage.clear();
+    return null;
   }
 }
